Will stop the party whenever a player isnt ready
In order to have the game compatible with the JavaFX application
Whenever the next player isn't ready, for example when the player
didn't clicked yet, the game will be stoped. This same player will
have to make the game continue whenever he is ready.

For that isReady methods has been added to the IO interface and
the player. This method has to be call before any call to players.